+++ screenController/solidColorProvider.cpp
@@ -1,12 +1,12 @@
 #include "solidColorProvider.h"
 
-double color[3]; 
-SolidColorProvider::SolidColorProvider(double* inputColor){
+uint8_t color[3]; 
+SolidColorProvider::SolidColorProvider(uint8_t* inputColor){
   color[0] = inputColor[0]; //R
   color[1] = inputColor[1]; //G
   color[2] = inputColor[2]; //B
 };
 
-double* SolidColorProvider::getColor(){
+uint8_t* SolidColorProvider::getColor(){
   return color;
 }
